Unify ingestion of precompiled GPU kernels in tests
We have quite a few tests that load kernels via `MultiKernelLoaderSpec`. Most of these tests load a kernel using `AddCudaPtxInMemory` from an embedded PTX string.

The same is not possible for ROCm because there is no equivalent abstraction layer to PTX. So for ROCm we load an embedded precompiled HSACO (fat) binary.

This change unifies both approaches by introducing a genrule that can extract a CUDA or ROCm fat binary from a `gpu_kernel_library` target.

Concretely it does the following:

1. Create new target `:gpu_test_kernels_fatbin` which complements `:gpu_test_kernels` and offers the same kernels as a fat binary (CUBIN/HSACO) in memory.
2. Extend `gpu_kernel_test` so that it test all kernel loading APIs `AddCudaPtxInMemory`, `AddCudaCubinInMemory`, `AddInProcessSymbol`.
3. Use the correct kernels names `AddI32` and `MulI32` in all tests (this matters since we don't load from PTX anymore)
4. Remove the old ROCm only kernel embedding approach which only worked for ROCm.

PiperOrigin-RevId: 661995313